chore(storybook): restructure storybook for maintainability
To keep the stories with their target components, and avoid lots of
long paths in the config.
