+++ .storybook/config.js
@@ -23,8 +23,7 @@ iconXHR.onload = function (e) {
 }
 
 function loadStories () {
-  require('../app/redux-app/components/stories/Button')
-  require('../app/redux-app/components/stories/Icon')
+  require('../app/redux-app/components/stories.js')
 }
 
 configure(loadStories, module)
