+++ app/redux-app/components/Button/story.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf, action } from '@kadira/storybook'
-import Button from '../Button'
+import Button from '.'
 
 storiesOf('Button', module)
   .add('plain', () => (
