+++ app/redux-app/components/Icon/index.js
