+++ app/redux-app/components/LogoLoader/index.js
