+++ app/redux-app/components/LogoLoader/story.js
@@ -1,0 +1,19 @@
+import React from 'react'
+import { storiesOf } from '@kadira/storybook'
+import LogoLoader from '.'
+
+storiesOf('LogoLoader', module)
+  .add('default', () => (
+    <LogoLoader inverted={false} loading={false}/>
+  ))
+  .add('inverted colour', () => (
+    <LogoLoader inverted={true} loading={false}/>
+  ))
+  .add('loading', () => (
+    <ul>
+      <li><LogoLoader inverted={false} loading={true}/>
+        inverted: false, loading: true</li>
+      <li><LogoLoader inverted={true} loading={true}/>
+        inverted: true, loading: true</li>
+    </ul>
+  ))
