+++ app/redux-app/components/stories.js
@@ -1,0 +1,9 @@
+/**
+ * Collection of component stories for React Storybook.
+ *
+ * See .storybook for main config
+ */
+
+require('./Button/story.js')
+require('./Icon/story.js')
+require('./LogoLoader/story.js')
