--- app/redux-app/components/stories/README.md
@@ -1,11 +1,0 @@
-# Component Stories
-
-These are "stories" for [react-storybook](https://github.com/kadirahq/react-storybook).
-
-Each describes some states for a component, allowing us to view the
-component in the storybook in any of those states. This is useful
-during development and manual testing of components.
-
-Run the storybook with `npm run storybook`
-
-Import new stories to the configuration in `<zanata-spa>/.storybook/config.js`
